separated music functions to its own file
